pyproject.toml
[project]
name = "youtubedownloadbot"
version = "0.1.0"
description = ""
authors = [
    {name = "sulgeese",email = "sul.geese@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp (>=2025.2.19,<2026.0.0)",
    "aiogram (>=3.19.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
